# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Entrepreneur Mitra Backend",
    "description": "Production-quality backend for SIH26092: AI-Driven Scheme Matching for Marginalized Entrepreneurs (Ministry of Social Justice and Empowerment - MoSJE). Connects conversational voice interview, deterministic rule evaluation, transparent ranking, financial affordability calculation, and geo-spatial partner routing.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/interviews/{conversation_id}/turn": {
      "post": {
        "tags": [
          "AI Interview"
        ],
        "summary": "Process Interview Turn",
        "description": "Process a speech/text utterance in the conversational AI interview.\nExtracts structured fields and returns the next targeted question.",
        "operationId": "process_interview_turn_api_v1_interviews__conversation_id__turn_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Conversation or Profile identifier",
              "title": "Conversation Id"
            },
            "description": "Conversation or Profile identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewTurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_InterviewTurnResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interview/start": {
      "post": {
        "tags": [
          "AI Interview"
        ],
        "summary": "Start Interview",
        "description": "Start an AI interview session.",
        "operationId": "start_interview_api_v1_interview_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interview/message": {
      "post": {
        "tags": [
          "AI Interview"
        ],
        "summary": "Post Interview Message",
        "description": "Shorthand alias for interview turn.",
        "operationId": "post_interview_message_api_v1_interview_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_InterviewTurnResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interview/confirm": {
      "post": {
        "tags": [
          "AI Interview"
        ],
        "summary": "Confirm Interview Fields",
        "description": "User confirmation of candidate fields extracted by AI.",
        "operationId": "confirm_interview_fields_api_v1_interview_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmFieldsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/eligibility/check": {
      "post": {
        "tags": [
          "Eligibility"
        ],
        "summary": "Check Eligibility",
        "description": "Evaluate deterministic scheme eligibility for a given entrepreneur profile.\nReturns traceable criteria evaluation with exact pass/fail reasons.",
        "operationId": "check_eligibility_api_v1_eligibility_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EligibilityCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_SchemeEligibilityResult__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eligibility/simulate": {
      "post": {
        "tags": [
          "Eligibility"
        ],
        "summary": "Simulate What If",
        "description": "Simulate hypothetical eligibility when attributes change (e.g. income, project cost).\nZero database mutation; clearly flagged as hypothetical.",
        "operationId": "simulate_what_if_api_v1_eligibility_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfSimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WhatIfResultOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches": {
      "post": {
        "tags": [
          "Matching"
        ],
        "summary": "Match Schemes",
        "description": "Match and rank government schemes for an entrepreneur profile\nusing the transparent 6-factor weighting model.",
        "operationId": "match_schemes_api_v1_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matching/results": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Get Matching Results",
        "description": "Get matched schemes by query parameter.",
        "operationId": "get_matching_results_api_v1_matching_results_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Entrepreneur Profile ID",
              "title": "Profile Id"
            },
            "description": "Entrepreneur Profile ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{scheme_id}/explanation": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Explain Scheme Match",
        "description": "Return structured rule trace explaining exactly WHY a scheme was matched or rejected.",
        "operationId": "explain_scheme_match_api_v1_matches__scheme_id__explanation_get",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MatchExplanationOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_InterviewTurnResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewTurnResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[InterviewTurnResponse]"
      },
...
      "ApiResponse_List_SchemeEligibilityResult__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SchemeEligibilityResult"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[SchemeEligibilityResult]]"
      },
...
      "ApiResponse_MatchExplanationOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchExplanationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[MatchExplanationOut]"
      },
      "ApiResponse_MatchResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[MatchResponse]"
      },
...
      "ApiResponse_WhatIfResultOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WhatIfResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[WhatIfResultOut]"
      },
      "ApiResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "ConfirmFieldsRequest": {
        "properties": {
          "confirmed_fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Confirmed Fields"
          }
        },
        "type": "object",
        "required": [
          "confirmed_fields"
        ],
        "title": "ConfirmFieldsRequest"
      },
...
      "EligibilityCheckRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "scheme_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme Ids"
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "EligibilityCheckRequest"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_message": {
            "type": "string",
            "title": "User Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "user_message"
        ],
        "title": "ErrorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewStartRequest": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          }
        },
        "type": "object",
        "title": "InterviewStartRequest"
      },
      "InterviewTurnRequest": {
        "properties": {
          "input_type": {
            "type": "string",
            "title": "Input Type",
            "default": "text"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "hi"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "InterviewTurnRequest"
      },
      "InterviewTurnResponse": {
        "properties": {
          "assistant_message": {
            "type": "string",
            "title": "Assistant Message"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "extracted_attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Attributes"
          },
          "extracted_fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Fields"
          },
          "next_required_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Required Field"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": true
          },
          "uncertainties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainties"
          },
          "profile_completeness_pct": {
            "type": "integer",
            "title": "Profile Completeness Pct",
            "default": 0
          },
          "conversation_state": {
            "type": "string",
            "title": "Conversation State",
            "default": "in_progress"
          }
        },
        "type": "object",
        "required": [
          "assistant_message",
          "reply",
          "extracted_attributes",
          "extracted_fields"
        ],
        "title": "InterviewTurnResponse"
      },
...
      "MatchRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "MatchRequest"
      },
...
      "MetaInfo": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "MetaInfo"
      },
...
      "SchemeEligibilityResult": {
        "properties": {
          "scheme_id": {
            "type": "string",
            "title": "Scheme Id"
          },
          "scheme_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionResult"
            },
            "type": "array",
            "title": "Criteria",
            "default": []
          },
          "matched_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Rules",
            "default": []
          },
          "failed_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed Rules",
            "default": []
          },
          "missing_information": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Information",
            "default": []
          },
          "passed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Passed",
            "default": []
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing",
            "default": []
          },
          "summary_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Explanation"
          }
        },
        "type": "object",
        "required": [
          "scheme_id",
          "status"
        ],
        "title": "SchemeEligibilityResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WhatIfResultOut": {
        "properties": {
          "is_hypothetical": {
            "type": "boolean",
            "title": "Is Hypothetical",
            "default": true
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "scheme_id": {
            "type": "string",
            "title": "Scheme Id"
          },
          "original": {
            "additionalProperties": true,
            "type": "object",
            "title": "Original"
          },
          "simulated": {
            "additionalProperties": true,
            "type": "object",
            "title": "Simulated"
          },
          "changed_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Rules"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "default": "This is a hypothetical simulation. Do not provide false information in actual applications."
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "scheme_id",
          "original",
          "simulated",
          "changed_rules",
          "explanation"
        ],
        "title": "WhatIfResultOut"
      },
      "WhatIfSimulateRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "scheme_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme Id"
          },
          "hypothetical_changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hypothetical Changes"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "hypothetical_changes"
        ],
        "title": "WhatIfSimulateRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}